refactor parser code and add unit tests
