--- tests/test_imSim.py
@@ -1,24 +1,0 @@
-"""
-Example unit tests for imSim package
-"""
-import unittest
-import desc.imsim
-
-class imSimTestCase(unittest.TestCase):
-    def setUp(self):
-        self.message = 'Hello, world'
-        
-    def tearDown(self):
-        pass
-
-    def test_run(self):
-        foo = desc.imsim.imSim(self.message)
-        self.assertEquals(foo.run(), self.message)
-
-    def test_failure(self):
-        self.assertRaises(TypeError, desc.imsim.imSim)
-        foo = desc.imsim.imSim(self.message)
-        self.assertRaises(RuntimeError, foo.run, True)
-
-if __name__ == '__main__':
-    unittest.main()
