+++ tests/tiny_instcat.txt
@@ -1,0 +1,28 @@
+rightascension 31.1133844
+declination -10.0970060
+mjd 59797.2854090
+altitude 43.6990272
+azimuth 73.7707957
+filter 2
+rotskypos 69.0922930
+dist2moon 145.1095257
+moonalt -11.1383568
+moondec -18.7702120
+moonphase 59.6288830
+moonra 230.9832941
+nsnap 2
+obshistid 161899
+rottelpos -53.5814660
+seed 161899
+seeing 0.7613760
+sunalt -59.1098785
+vistime 33.0000000
+object 1046817878020 31.2400746 -10.09365 29.3370237 starSED/phoSimMLT/lte033-4.5-1.0a+0.4.BT-Settl.spec.gz 0 0 0 0 0 0 point CCM 0.0635117705 3.1 none
+object 811883456516 31.2338048 -10.0578789 26.9267056 starSED/phoSimMLT/lte027-2.0-0.0a+0.0.BT-Settl.spec.gz 0 0 0 0 0 0 point CCM 0.0630700848 3.1 none
+object 1046816368644 31.2371416 -10.0554117 28.7153047 starSED/phoSimMLT/lte027-2.0-0.0a+0.0.BT-Settl.spec.gz 0 0 0 0 0 0 point CCM 0.0630379733 3.1 none
+object 1046816356356 31.2305508 -10.0620637 26.7932237 starSED/phoSimMLT/lte033-4.5-1.0a+0.4.BT-Settl.spec.gz 0 0 0 0 0 0 point CCM 0.0631214889 3.1 none
+object 1046816650244 31.2225696 -10.0784635 29.2177099 starSED/phoSimMLT/lte032-4.5-1.0a+0.4.BT-Settl.spec.gz 0 0 0 0 0 0 point CCM 0.0632916884 3.1 none
+object 275781785604 31.2570189 -10.0836452 23.454705 starSED/phoSimMLT/lte027-2.0-0.0a+0.0.BT-Settl.spec.gz 0 0 0 0 0 0 point CCM 0.0638548764 3.1 none
+object 956090372100 31.2608206 -10.0836426 24.218305 starSED/phoSimMLT/lte027-2.0-0.0a+0.0.BT-Settl.spec.gz 0 0 0 0 0 0 point CCM 0.0651282621 3.1 none
+object 956090392580 31.258544 -10.0948714 24.6497053 starSED/phoSimMLT/lte027-2.0-0.0a+0.0.BT-Settl.spec.gz 0 0 0 0 0 0 point CCM 0.0639515271 3.1 none
+object 811883374596 31.2499807 -10.032142 23.1621553 starSED/phoSimMLT/lte031-4.5-1.0a+0.4.BT-Settl.spec.gz 0 0 0 0 0 0 point CCM 0.0651531353 3.1 none
