+++ ups/imsim.table
@@ -1,3 +1,6 @@
+setupRequired(sims_catalogs_measures)
+setupRequired(sims_catalogs_generation)
+
 envPrepend(PATH, ${PRODUCT_DIR}/bin)
 
 envPrepend(LD_LIBRARY_PATH, ${PRODUCT_DIR}/lib)
